Cleanup a bit
