+++ Rakefile
@@ -1,8 +1,10 @@
+require "bundler/gem_tasks"
 require 'rake'
 require 'rake/testtask'
 
 Rake::TestTask.new do |t|
-  t.test_files = %w(midiskirt_test.rb)
+  t.libs << 'test'
+  t.test_files = Dir["test/**/*_test.rb"]
 end
 
 task :default => :test
