+++ lib/midiskirt/version.rb
@@ -1,0 +1,3 @@
+class Midiskirt
+  VERSION = "1.2.0"
+end
